+++ preprocess.php
@@ -23,10 +23,6 @@
 	});
 	
 	set_exception_handler(function($exception) {
-		if(!error_reporting()) {
-			return;
-		}
-		
 		if(php_sapi_name() == 'cli') {
 			die($exception);
 		}
@@ -41,18 +37,17 @@
 		
 		$stack .= '</ol>'."\n";
 		
-		echo '<pre style="">';
-		echo '<h1 style="margin:0px;">Uncaught Exception: '.get_class($exception).'</h1><br>';
-		echo '<strong>Code: </strong>'.$exception->getCode().'<br>';
-		echo '<strong>File: </strong>'.$exception->getFile().'<br>';
-		echo '<strong>Line: </strong>'.$exception->getLine().'<br>';
-		echo '<strong>Message: </strong>'.$exception->getMessage().'<br>';
-		echo '<strong>Stacktrace: </strong><br>'.$stack;
-		echo '</pre>';
+		print '<pre style="">';
+		print '<h1 style="margin:0px;">Uncaught Exception: '.get_class($exception).'</h1><br>';
+		print '<strong>Code: </strong>'.$exception->getCode().'<br>';
+		print '<strong>File: </strong>'.$exception->getFile().'<br>';
+		print '<strong>Line: </strong>'.$exception->getLine().'<br>';
+		print '<strong>Message: </strong>'.$exception->getMessage().'<br>';
+		print '<strong>Stacktrace: </strong><br>'.$stack;
+		print '</pre>';
 	});
 	
 	set_error_handler(function($errno, $errstr, $errfile, $errline, $errcontext) {
-		
 		if(!(error_reporting() & $errno)) {
 			return;
 		}
@@ -80,21 +75,21 @@
 		array_pop($trace);
 		
 		if(php_sapi_name() == 'cli') {
-			echo 'Backtrace from ' . $type . ' \'' . $errstr . '\' at ' . $errfile . ' ' . $errline . ':' . "\n";
+			print 'Backtrace from ' . $type . ' \'' . $errstr . '\' at ' . $errfile . ' ' . $errline . ':' . "\n";
 			foreach($trace as $item) {
-				echo '  ' . (isset($item['file']) ? $item['file'] : '<unknown file>') . ' line ' . (isset($item['line']) ? $item['line'] : '<unknown line>') . ' calling ' . $item['function'] . '()' . "\n";
+				print '  ' . (isset($item['file']) ? $item['file'] : '<unknown file>') . ' line ' . (isset($item['line']) ? $item['line'] : '<unknown line>') . ' calling ' . $item['function'] . '()' . "\n";
 			}
 		} else {
-			echo '<pre class="alert alert-danger" style="">' . "\n";
-			echo '<p style="line-height:10px; margin:0px;">Backtrace from ' . $type . ' \'' . $errstr . '\' at ' . $errfile . ' ' . $errline . ':' . "\n".'</p>';
-			echo '  <ol style="margin-top:0px; line-height:10px;">' . "\n";
+			print '<pre class="alert alert-danger" style="">' . "\n";
+			print '<p style="line-height:10px; margin:0px;">Backtrace from ' . $type . ' \'' . $errstr . '\' at ' . $errfile . ' ' . $errline . ':' . "\n".'</p>';
+			print '  <ol style="margin-top:0px; line-height:10px;">' . "\n";
 			
 			foreach($trace as $item) {
-				echo '<li>' . (isset($item['file']) ? $item['file'] : '<unknown file>') . ' line ' . (isset($item['line']) ? $item['line'] : '<unknown line>') . ' calling ' . $item['function'] . '()</li>' . "\n";
+				print '<li>' . (isset($item['file']) ? $item['file'] : '<unknown file>') . ' line ' . (isset($item['line']) ? $item['line'] : '<unknown line>') . ' calling ' . $item['function'] . '()</li>' . "\n";
 			}
 			
-			echo '  </ol>' . "\n";
-			echo '</pre>' . "\n";
+			print '  </ol>' . "\n";
+			print '</pre>' . "\n";
 		}
 		
 		if(ini_get('log_errors')) {