+++ resources/themes/default/footer.tpl.php
@@ -1,2 +1,5 @@
+	<?php foreach($app->document->getJavascript() as $script): ?>
+		<script type="text/javascript" src="<?php print $script; ?>"></script>
+	<?php endforeach; ?>
 	</body>
 </html>