+++ resources/themes/default/functions.php
@@ -1,0 +1,5 @@
+<?php
+	$app->document->addJavascript( Helper::url("/public/stylesheets/jquery-1.11.1.min.js") );
+	
+	$app->document->addStylesheet( Helper::url("/public/stylesheets/screen.css") );
+?>